backend/requirements.txt
Django>=5.2.6
djangorestframework>=3.16.1
django-cors-headers>=4.9.0
dj-database-url>=3.0.1
psycopg2-binary>=2.9.10
python-dotenv>=1.1.1
PyJWT>=2.10.1
supabase>=2.0
